Move upgrade code to its own module
