+++ collect_app/src/main/java/org/odk/collect/android/application/initialization/ApplicationInitializer.kt
@@ -15,12 +15,12 @@ import org.javarosa.xform.parse.XFormParser
 import org.odk.collect.analytics.Analytics
 import org.odk.collect.android.BuildConfig
 import org.odk.collect.android.application.Collect
-import org.odk.collect.android.application.initialization.upgrade.AppUpgrader
 import org.odk.collect.android.geo.MapboxUtils
 import org.odk.collect.android.logic.PropertyManager
 import org.odk.collect.android.logic.actions.setgeopoint.CollectSetGeopointActionHandler
 import org.odk.collect.android.preferences.source.SettingsProvider
 import org.odk.collect.projects.ProjectsRepository
+import org.odk.collect.upgrade.AppUpgrader
 import org.odk.collect.utilities.UserAgentProvider
 import org.osmdroid.config.Configuration
 import timber.log.Timber
