+++ collect_app/src/main/java/org/odk/collect/android/application/initialization/upgrade/BeforeProjectsInstallDetector.kt
@@ -3,6 +3,7 @@ package org.odk.collect.android.application.initialization.upgrade
 import android.content.Context
 import androidx.preference.PreferenceManager
 import org.odk.collect.android.utilities.FileUtils
+import org.odk.collect.upgrade.InstallDetector
 import java.io.File
 
 class BeforeProjectsInstallDetector(private val context: Context) : InstallDetector {
