+++ settings.gradle
@@ -17,4 +17,5 @@ include ':location'
 include ':geo'
 include ':errors'
 include ':externalapp'
+include ':upgrade'
 include ':permissions'
